introduce-output-streamer-move: Move output streaming to SchedulerOutputStreamer
Mechanical cut + paste for the ``introduce-output-streamer`` mech move.

Cut the stream @staticmethods (after prep) from
``SchedulerOutputProcessorMixin`` and paste them into
``SchedulerOutputStreamer`` body in
``scheduler_components/output_streamer.py``.

Drop ``@staticmethod`` decorators; simplify
``self: "SchedulerOutputStreamer"`` → bare ``self``. Method bodies
byte-identical (all Callable substitutions + privacy flips happened in
prep so this step is pure mechanical cut+paste).

All callers updated:
  ``self.<m>(self.output_streamer, ...)`` →
  ``self.output_streamer.<m>(...)``
(pure prefix transformation).

Refactor chain ID: introduce-output-streamer-move
